requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
greenlet==3.0.3
kiwisolver==1.4.5
matplotlib==3.9.0
mysql-connector-python==9.0.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pandasql==0.7.3
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
